Use ingest info object cache for all ND scrapers (Recidiviz/recidiviz-data#1912)
Update the us_nd_scraper so that all files use the ingest_info_object_cache, which allows us to properly handle dummy objects and avoid trees with objects with duplicate ids, in order to support entity matching properly.

A few main changes:

* Update _convert_to_supervision_sentence code to properly update the cache to remove the old placeholder incarceration sentence and re-register as a supervision sentence
* Pushed some logic into _get_object_to_set so it always returns a nonnull object value. In that function, added logic to lookup the object in the object cache and return that immediately if we find it.
* Handled case where you find that you want to create a second instance of a child object on a dummy parent that only supports singular child objects of that type (changes inside of _get_cached_parent)

The test changes are mostly just a result of us merging dummy objects where it is now appropriate.

GitOrigin-RevId: fd2fd8d780428fa6c438df82d2a71e05b542c6a5