allow replacement of singular standalone entities (Recidiviz/recidiviz-data#2590)
Allow replacement of singular entities in the person tree that are standalone (just StateAgent for now). Because these entities can exist separate from a StatePerson, we don't need to worry about orphaning them.

The lack of this functionality is causing entity matching errors when we update parole officers in docstars_offenders (functionality added in Recidiviz/recidiviz-data#2558)

GitOrigin-RevId: ef48333099e0731843dc1f8ade5c0f9b303acfd2