[Calculate] Properly filter MO people out of supervision population when they were not on supervision at any point in the month (Recidiviz/recidiviz-data#3096)
With this PR, we will now be properly omitting people from the supervision population in MO when they are not on supervision at any point in the month. And also ensures we don't count people towards the end of month population if they don't have a supervision period overlapping with the end of the month.

In the spirit of Tech Debt Month, did a bit of refactoring to put more general logic in helper classes that can be unit tested individually.

In this PR:

* Added new classes for representing time ranges and the difference between ranges in time_range_utils.py

* Tests for time_range_utils.py

* Moved all incarceration period indexing to a new IncarcerationPeriodIndex class in incarceration_period_index.py. This includes some new functionality, including the month_to_overlapping_incarceration_period index and the is_fully_incarcerated_for_range function

* Tests for incarceration_period_index.py, including lots of new ones and some that I moved out of the supervision identifier_test.py file.

* Updated the supervision pipeline identifier to use the new IncarcerationPeriodIndex instead of passing a bunch of maps around.

* Created a generalized supervision_period_counts_towards_supervision_population_in_date_range function, then updated the identifier to use that function in a number of places to consolidate logic.

GitOrigin-RevId: 8f4f161e50b8b3c6e3364f362e7ba79f027e8594