Fix Postgres -> BQ export (Recidiviz/recidiviz-data#3203)
I was trying to get the num_rows out of a TableReference instead of a Table. This fixes to get the actual table for the log statement, and also does some cleanup to limit the number of places we're instantiating a BigQueryClientImpl, passing one around instead.

I was able to successfully run a local export with these changes.

GitOrigin-RevId: dea4d3f660e8ad485ad50f3adb1d5683a7e3f84d