[SQL preprocessing] Use PostgresDirectIngestMetadataManager in tests (Recidiviz/recidiviz-data#3239)
Now that we're using a SQLAlchemy postgres implementation for the direct ingest metadata manager, there's no need to mock the whole class - we can just use an in-memory version of the DB and get much more realistic behavior (tests will even make sure we're not violating constraints, etc).

Switching to the Postgres implementation was a little more complicated than planned, because some of the tests in gcsfs_direct_ingest_controller_test.py run on background threads and our current implementation of the in-memory sqlite DB did not configure the DB to allow for multi-threaded shared acces.

With some help from StackOverflow, I have now updated the test helper to share the cache.

I also updated every test to call a teardown that makes sure we clear the cached DB for the next test.

Work towards Recidiviz/recidiviz-data#3020

GitOrigin-RevId: 9bc3c930f4a6ed0f75c2971a30b2f4b0c7ac3245