Use an on-disk Postgres DB in tests that use multiple threads (Recidiviz/recidiviz-data#3266)
I've been running up against a bunch of issues with getting sqlite databases to work when running on multiple threads. This PR adds functionality to our travis build to allow us to connect to an on-disk postgres database for use in tests that need to access one of our databases from multiple threads.

These new tests helpers work locally as well as in Travis.

Work towards Recidiviz/recidiviz-data#3020

GitOrigin-RevId: fa015613b6f56fcf024c963468c728f02398da4c