[PA ingest] Ingest all external ids ever associated with a person (Recidiviz/recidiviz-data#3419)
Wrote a query that assigns a master state id to all associated control_number, state_id, and parole_number values. This query can be used in future queries, but I also developed a new ingest view that pulls in all ids for a given person at the very beginning so entity matching has as much info as possible at the start.

Note: I started building this query to be recursive because SQL apparently has a WITH RECURSIVE statement that allows you to do recursive subqueries, but found out too late that BQ doesn't support this, so this is two iterations of recursion which turns out to be good enough.

Closes Recidiviz/recidiviz-data#3407

GitOrigin-RevId: d6ed5116a568fe4a15bdfab00dbf7a8329b4f1e3