Face to face contacts (Recidiviz/recidiviz-data#3415)
This PR adds a contact_frequency_sufficient field to the SupervisionComplianceMetrics which measures whether a supervision officer has had enough face-to-face contacts with the person on supervision within a time period that satisfies the state's minimum contact frequency requirements. Implemented for US_ID only for use in the PO Report.

GitOrigin-RevId: b22163f4c43c49719c32a9b4c7295d0f252c8bfe